refactor(main): split commands into koda/commands/ package
main.py had grown to ~1350 lines with every Typer command in one file,
a constant source of merge conflicts. Split it into focused modules:

- runtime.py: shared CLI runtime with no Typer commands — lazy
  get_config/get_db (so importing the CLI has no side effects), init_db,
  version_callback, and cross-command helpers (resolve_ref, emit_raw,
  _apply_vars, _strip_inline_comment(s), _read_stdin_refs,
  _validate_list_columns).
- commands/memo.py: add, remove, copy, edit, list, show, raw, tag.
- commands/index.py: move, swap, shift, compact.
- commands/exec.py: exec, pick, _run_pick_action.
- commands/git.py: push, pull.
- commands/config.py: the config_app subcommand group.
- main.py: now ~110 lines — ALIASES, KodaGroup, the root app + callback,
  and the bottom imports that register every command.

Import wiring: command modules import `app` (and RESERVED_SHORTCUTS)
from main, so main imports them at the bottom after `app` is defined.
commands/config.py does NOT import main (imported at top for
app.add_typer). commands/exec.py binds `from . import memo` as a module
and calls memo.edit() at runtime to avoid a load-time cycle.

No behavior change: full CLI smoke-tested (aliases, default-cmd routing,
add/list/show/tag/exec, config). Tests that referenced moved symbols are
updated to the new module locations; the import-has-no-side-effects test
now targets runtime. CLAUDE.md architecture section updated.

Closes #82
