+++ src/koda/commands/exec.py
@@ -1,0 +1,144 @@
+"""Execution and interactive-selection commands: exec, pick."""
+
+import os
+import sys
+
+import typer
+
+from ..cli_utils import exit_error
+from ..cmd_helpers.display import print_memo as _print_memo
+from ..cmd_helpers.interactive import pick_candidates, pick_with_fzf, resolve_pick_action
+from ..config import ConfigManager, ValidationError
+from ..main import app
+from ..runtime import (
+    _apply_vars,
+    _read_stdin_refs,
+    emit_raw,
+    get_config,
+    get_db,
+    init_db,
+    resolve_ref,
+)
+from . import memo  # bound as a module to avoid a circular import at load time
+
+
+def _run_pick_action(action: str, ref: str) -> None:
+    if action == "raw":
+        emit_raw(ref)
+        return
+    if action == "show":
+        init_db()
+        row = resolve_ref(ref)
+        _print_memo(row.uid, row.idx, row.shortcut, row.content, row.tags, row.created_at)
+        return
+    if action == "edit":
+        memo.edit(ref)
+        return
+    if action == "exec":
+        exec_memo(ref, None)
+        return
+    exit_error(f"Unsupported pick action: {action}")
+
+
+@app.command(name="exec")
+def exec_memo(
+    ref: str | None = typer.Argument(
+        None, help="Entry index or shortcut to execute (default: latest)."
+    ),
+    vars: list[str] | None = typer.Option(
+        None,
+        "--var",
+        "-V",
+        help=(
+            "Variable substitution. Named: KEY=VALUE → replaces ${KEY}. "
+            "Positional: VALUE → replaces $1,$2,... in order. "
+            'Comma-separate multiple values; use "..." to include spaces or commas. '
+            "Examples: -V 'localhost,5432' -V 'name=prod' -V '\"hello world\",\"foo,bar\"'"
+        ),
+    ),
+):
+    """Execute the memo body as a shell command. Alias: `koda x`.
+
+    When no argument is given, this command also accepts one ref from stdin.
+    """
+    if ref is None:
+        stdin_refs = _read_stdin_refs()
+        if len(stdin_refs) > 1:
+            exit_error("ex accepts one ref from stdin. Got multiple values.")
+        if stdin_refs:
+            ref = stdin_refs[0]
+
+    init_db()
+    row = resolve_ref(ref)
+    content = _apply_vars(row.content.strip() if row.content else "", vars)
+    shell = get_config().exec_shell
+    try:
+        ConfigManager.validate("exec.shell", shell)
+    except ValidationError:
+        exit_error(
+            f"Refusing to exec: exec.shell {shell!r} is not allowed "
+            f"({ConfigManager.error_message('exec.shell')})."
+        )
+    os.execvp(shell, [shell, "-c", content])
+
+
+@app.command()
+def pick(
+    query: str | None = typer.Option(None, "--query", "-q", help="Substring match on memo body."),
+    tag: str | None = typer.Option(None, "--tag", "-t", help="Substring match on tags."),
+    exclude_tag: str | None = typer.Option(
+        None, "--exclude-tag", "-T", help="Exclude entries whose tags include this substring."
+    ),
+    shortcuts_only: bool = typer.Option(
+        False, "--shortcuts", "-S", help="Show only entries that have a shortcut."
+    ),
+    sort_by: str | None = typer.Option(
+        None,
+        "--sort-by",
+        case_sensitive=False,
+        help=(
+            "Sort column: id, idx, uid, tags, content, created_at, modified_at, shortcut. "
+            "[config: list.sort_by]"
+        ),
+    ),
+    desc: bool | None = typer.Option(
+        None,
+        "--desc/--asc",
+        help="Sort order. [config: list.desc]",
+    ),
+    print_id: bool = typer.Option(
+        False, "--print-id", "-p", help="Print selected IDX and exit without running a command."
+    ),
+    edit_mode: bool = typer.Option(False, "--edit", "-e", help="Open selected entry in editor."),
+    exec_mode: bool = typer.Option(False, "--exec", "-x", help="Execute selected entry."),
+    raw_mode: bool = typer.Option(False, "--raw", "-r", help="Print selected entry body."),
+    show_mode: bool = typer.Option(
+        False, "--show", "-s", help="Show selected entry with metadata."
+    ),
+):
+    """Pick an entry with fzf, then run an action (or print IDX). Alias: `koda p`."""
+    if print_id and (edit_mode or exec_mode or raw_mode or show_mode):
+        exit_error("--print-id/-p cannot be combined with action flags.")
+
+    action: str | None = (
+        None
+        if print_id
+        else resolve_pick_action(get_config(), edit_mode, exec_mode, raw_mode, show_mode, print_id)
+    )
+
+    init_db()
+    candidates = pick_candidates(
+        get_db(), get_config(), query, tag, exclude_tag, shortcuts_only, sort_by, desc
+    )
+    if not candidates:
+        exit_error("No entries found.", style="yellow")
+
+    selected_ref = pick_with_fzf(candidates)
+    if selected_ref is None:
+        raise typer.Exit(code=0)
+
+    if print_id:
+        sys.stdout.write(selected_ref + "\n")
+        return
+
+    _run_pick_action(action, selected_ref)
